Move all source to the root folder
Honestly, this library should not grow too much. Its idea is to be
small, so if we need subdirectories then the code needs to be split into
multiple repos.

Also, it is cooler to break this assumption of all source code requiring
folders and stuff.

Also also, it's more productive in VIM/Emacs because one doesn't need to
spend time searching for code.