removing TIMER
This is more headache than solution. It was nice to have a 20tps timer
for certain game loops, but it is proving to be quite a headache.

First, we might want to tune the timer or have more than one in the same
app. Can't be done in casein.

Then, iOS requires all timers to pause when entering background. This is
a nasty solution for a single platform.

Therefore, it is better to implement an API for timers than trying to
support it on casein.